fix compass rotation

+++ domain/src/main/java/com/isaevapps/domain/model/CompassData.kt
@@ -1,5 +1,5 @@
 package com.isaevapps.domain.model
 
 data class CompassData(
-    val azimuth: Float
+    val azimuth: Int
 )
